adding testing and starting to add installer

+++ src/alz_mri_cnn/front_end.py
@@ -3,7 +3,9 @@ import pathlib
 import pickle
 import random
 import shutil
+import signal
 import typing
+from multiprocessing import Process
 
 import cv2
 import keras
@@ -11,15 +13,31 @@ import numpy as np
 from flask import Flask, render_template, request
 from werkzeug.utils import secure_filename
 
-IMG_SIZE = (128, 128)
-# IMG_SIZE = (128//2, 128//2)
+from alz_mri_cnn.utils import IMG_SIZE, RUNNING_DIR
 
-RUNNING_DIR = "/tmp/alz_mri_cnn/"
+CATEGORIES = None
+
 app = Flask(__name__)
-
 
 model = None
 model_accuracy = None
+
+
+@app.route("/shutdown", methods=["GET"])
+def shutdown():
+    global server
+    print(f" ==   run server terminate: {server}")
+    os.kill(server.pid, signal.SIGKILL)
+    print(f" ==   server terminated: {server}")
+    return "Server shutting down..."
+
+
+def start_local_server():
+    global server
+    server = Process(target=app.run)
+    server.start()
+    print(f" ==   server set: {server}")
+    return server
 
 
 def get_model() -> keras.Model:
@@ -30,7 +48,7 @@ def get_model() -> keras.Model:
 
         def find_model_in_dir(dir):
             for k in os.listdir(dir):
-                if '.keras' in k:
+                if ".keras" in k:
                     found_models.append(os.path.join(dir, k))
 
         best_path_1 = os.path.join(RUNNING_DIR, "models")
@@ -38,31 +56,38 @@ def get_model() -> keras.Model:
 
         best = None
         for model in found_models:
-            acc = int(
-                model[model.find("%") - 2: model.find("%")].replace("_", "")
-            )
-            if best is None or acc > best[0]:   # type: ignore
+            acc = int(model[model.find("%") - 2: model.find("%")].replace("_", ""))
+            if best is None or acc > best[0]:  # type: ignore
                 best = (acc, model)
 
-        model_accuracy, model_name = best   # type: ignore
+        model_accuracy, model_name = best  # type: ignore
 
         print(f"    loading model: {model_name} with accuracy: {model_accuracy}%")
         model = keras.models.load_model(model_name)
     return model
 
 
-CATEGORIES = None
+def get_categories(categories=None):
+    global CATEGORIES
+    if categories and not CATEGORIES:
+        CATEGORIES = categories
 
-
-def get_categories(CATEGORIES=CATEGORIES):
     if not CATEGORIES:
-        category_file = os.path.join(RUNNING_DIR, 'data', 'categories')
+        category_file = os.path.join(RUNNING_DIR, "data", "categories")
         if pathlib.Path(category_file).exists():
-            file = open(category_file, 'rb')
+            file = open(category_file, "rb")
             CATEGORIES = pickle.load(file)
 
         assert CATEGORIES
     return CATEGORIES
+
+
+def predict_on_request(request):
+    class_name = request.form.get("impairment_val")
+    print(f"    searching for random image of class_name: {class_name}")
+    image = get_random_image_of_class(class_name)
+    prediction, confidence = predict_image(image)
+    return image, prediction, confidence
 
 
 @app.route("/", methods=["POST", "GET"])
@@ -72,9 +97,7 @@ def on_start():
         of that class and the prediction the model gave.
     """
     if request.method == "POST":
-        class_name = request.form.get("impairment_val")
-        print(f'    searching for random image of class_name: {class_name}')
-        image, prediction, confidence = get_random_of_class(class_name)
+        image, prediction, confidence = predict_on_request(request)
     elif request.method == "GET":
         return render_template("index.html")
 
@@ -82,27 +105,31 @@ def on_start():
         get_model()
 
     # Clear out the static dir of all previous entries (jpgs)
-    for p in os.listdir("static"):
-        if ".jpg" in p:
-            try:
-                os.remove(f"static/{p}")
-            except Exception as e:
-                print(f"encountered issue when removing image {p} from static dir: {e}")
-    shutil.copy(image, "static")
+    if image:
+        for p in os.listdir("static"):
+            if ".jpg" in p:
+                try:
+                    os.remove(f"static/{p}")
+                except Exception as e:
+                    print(
+                        f"encountered issue when removing image {p} from static dir: {e}"
+                    )
+        shutil.copy(image, "static")
 
-    # Render the image now that it is in the static dir
-    index = image.rindex("/")
-    img_location = os.path.join("static", image[index + 1: len(image)])
-    return render_template(
-        "index.html",
-        model_accuracy=model_accuracy,
-        result=prediction,
-        image=img_location,
-        confidence=confidence,
-    )
+        # Render the image now that it is in the static dir
+        index = image.rindex("/")
+        img_location = os.path.join("static", image[index + 1: len(image)])
+        return render_template(
+            "index.html",
+            model_accuracy=model_accuracy,
+            result=prediction,
+            image=img_location,
+            confidence=confidence,
+        )
+    return render_template("index.html")
 
 
-def get_random_of_class(chosen_class):
+def get_random_image_of_class(chosen_class):
     """
     With the specified chosen_class, find a random image and get a prediction of that image from the model.
     """
@@ -115,8 +142,7 @@ def get_random_of_class(chosen_class):
             images = os.listdir(os.path.join(dir, path))
     assert images
     image = random.choice(images)
-
-    return predict_image(os.path.join(dir, get_categories()[index], image))
+    return os.path.join(dir, get_categories()[index], image)
 
 
 def predict_image(path):
@@ -132,7 +158,7 @@ def predict_image(path):
     x_data_reshape = np.reshape(x_data, (1, IMG_SIZE[0], IMG_SIZE[1], 3))
     probabilities = get_model().predict(x_data_reshape)
     max = np.argmax(probabilities)
-    return (path, get_categories()[max], int(probabilities[0][max] * 100))
+    return (get_categories()[max], int(probabilities[0][max] * 100))
 
 
 @app.route("/predict", methods=["POST"])
@@ -158,11 +184,12 @@ def predict():
 
 
 if __name__ == "__main__":
+    get_categories()
+
+    global server
+    server = start_local_server()
+
     # frontend must be run from src/alz_mri_cnn dir
     new_dir = os.path.dirname(os.path.abspath(__file__))
     os.chdir(new_dir)
-    print(f'cd into dir: {new_dir}')
-
-    get_categories()
-
-    app.run(debug=True)
+    print(f"cd into dir: {new_dir}")
